[BUGFIX] Пофиксил баг из отзывов гугл плей маркета: маркер на карте отображался на неправильном месте при загрузке погоды, а был сдвинут. Это было связано с тем, что брались координаты местоположения от апи, и они очень неточные, 2 числа после запятой. Пришлось прокидывать полные координаты с карты

+++ app/src/main/java/com/desuzed/everyweather/data/mapper/location/UserLatLngMapper.kt
@@ -6,6 +6,6 @@ import com.desuzed.everyweather.domain.model.location.UserLatLng
 
 class UserLatLngMapper : EntityMapper<Location, UserLatLng> {
     override fun mapFromEntity(entity: Location): UserLatLng {
-        return UserLatLng(entity.latitude.toFloat(), entity.longitude.toFloat(), entity.time)
+        return UserLatLng(entity.latitude, entity.longitude, entity.time)
     }
 }
